% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify_params.R
\name{ageify_params}
\alias{ageify_params}
\title{Add age-based heterogeneity to select parameters}
\usage{
ageify_params(values, disease, param_names)
}
\arguments{
\item{values}{(list, named) Values list}

\item{disease}{(character) Disease short name}

\item{param_names}{(character) list of parameters to adjust}
}
\value{
A list matching the output format of \code{\link[=update_values]{update_values()}}
}
\description{
Add age-based heterogeneity to select parameters
}
