% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_transmissibility.R
\name{calculate_rho}
\alias{calculate_rho}
\title{Calculate spectral radius of a contact matrix}
\usage{
calculate_rho(values)
}
\arguments{
\item{values}{(list, named) Values list}
}
\value{
Spectral radius of the contact matrix (numeric)
}
\description{
Calculate spectral radius of a contact matrix
}
