% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_transmissibility.R
\name{calculate_transmissibility}
\alias{calculate_transmissibility}
\title{Calculate transmissibility from other model parameters from values list}
\usage{
calculate_transmissibility(values)
}
\arguments{
\item{values}{(list, named) Values list}
}
\value{
Transmissibility value (numeric)
}
\description{
Uses R0 derived for the hosp model; see https://gitlab.cscscience.ca/mhd/firm/EPACmodel_PandPrep
}
