% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify_params.R
\name{get_age_mult}
\alias{get_age_mult}
\title{Get age-based multipliers for parameters}
\usage{
get_age_mult(disease)
}
\arguments{
\item{disease}{(character) Disease short name}
}
\value{
Age-based multipliers for parameters (numeric)
}
\description{
Get age-based multipliers for parameters
}
