% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_transmissibility.R
\name{transmissibility_formula}
\alias{transmissibility_formula}
\title{Transmissiblity formula using model parameters}
\usage{
transmissibility_formula(
  rho,
  pop,
  R0,
  prop_hosp,
  prop_nonhosp_death,
  days_infectious_I_R,
  days_infectious_I_A,
  days_infectious_I_D
)
}
\arguments{
\item{rho}{(numeric) Spectral radius of the contact matrix}

\item{pop}{(numeric) Population by age group}

\item{R0}{(numeric) Average number of secondary cases from an index case in a wholly susceptible population}

\item{prop_hosp}{(numeric) Proportion of infections that are hospitalized}

\item{prop_nonhosp_death}{(numeric) Proportion of infections that are not hospitalized but are fatal}

\item{days_infectious_I_R}{(numeric) Infectious period for infections that simply recover}

\item{days_infectious_I_A}{(numeric) Infectious period for infections that are hospitalized}

\item{days_infectious_I_D}{(numeric) Infectious period for infections that are fatal (outside of hospital)}
}
\value{
Transmissibility value (numeric)
}
\description{
Transmissiblity formula using model parameters
}
