% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_flow.R
\name{update_flow}
\alias{update_flow}
\title{Update flow in values list}
\usage{
update_flow(values)
}
\arguments{
\item{values}{(list) List of value for input into the model}
}
\value{
Values (list)
}
\description{
Update flow in values list
}
