% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_values.R
\name{update_values}
\alias{update_values}
\title{Update values list with new parameters}
\usage{
update_values(
  params,
  model.name,
  state,
  pop,
  age_param_names = NULL,
  disease = NULL
)
}
\arguments{
\item{params}{(named vector) New parameters to update}

\item{model.name}{(character) Short name of model in the EPACmodel package}

\item{state}{(named vector) Initial state vector}

\item{pop}{(tibble) table of population by age group}

\item{age_param_names}{(character) Names of parameters to change by age. If NULL, don't update parameters by age.}

\item{disease}{(character) Short name of disease for which to update parameters by age. Must match an option available in \link{ageify_params}}
}
\value{
Values (list)
}
\description{
Update values list with new parameters
}
